aggiornamento alla versione funzionante

+++ Firmware Sistema/Sistema Pylon Rev2 18-07/Palo 2e3/RXROSSO_Palo2-3_Luci/RXROSSO_Palo2-3_Luci.ino
@@ -25,8 +25,8 @@ int lampeggianteP2 = 0;
 int lampeggianteP3 = 0;
 int onP2 = 0;
 int onP3 = 0;
-int lampmillisP2;
-int lampmillisP3;
+unsigned long lampmillisP2;
+unsigned long lampmillisP3;
 
 //Inizio Configurazioni Ricevente
 String Race = "3000";
